+++ projects/core/src/checkout/facade/checkout.service.spec.ts
@@ -1,9 +1,8 @@
 import { inject, TestBed } from '@angular/core/testing';
 import { Store, StoreModule } from '@ngrx/store';
 import { CartDataService } from '@spartacus/core';
-import { Address, AddressValidation } from '../../model/address.model';
-import { CardType, Cart, PaymentDetails } from '../../model/cart.model';
-import { DeliveryMode, Order } from '../../model/order.model';
+import { Cart } from '../../model/cart.model';
+import { Order } from '../../model/order.model';
 import * as fromCheckout from '../store/index';
 import { CheckoutService } from './checkout.service';
 
@@ -13,20 +12,6 @@ describe('CheckoutService', () => {
   let store: Store<fromCheckout.CheckoutState>;
   const userId = 'testUserId';
   const cart: Cart = { code: 'testCartId', guid: 'testGuid' };
-
-  const paymentDetails: PaymentDetails = {
-    id: 'mockPaymentDetails',
-  };
-
-  const address: Address = {
-    firstName: 'John',
-    lastName: 'Doe',
-    titleCode: 'mr',
-    line1: 'Toyosaki 2 create on cart',
-    town: 'Montreal',
-    postalCode: 'L6M1P9',
-    country: { isocode: 'CA' },
-  };
 
   beforeEach(() => {
     TestBed.configureTestingModule({
@@ -51,115 +36,6 @@ describe('CheckoutService', () => {
     }
   ));
 
-  it('should be able to get supported delivery modes', () => {
-    store.dispatch(
-      new fromCheckout.LoadSupportedDeliveryModesSuccess([
-        { code: 'mode1' },
-        { code: 'mode2' },
-      ])
-    );
-
-    let deliveryModes: DeliveryMode[];
-    service
-      .getSupportedDeliveryModes()
-      .subscribe(data => {
-        deliveryModes = data;
-      })
-      .unsubscribe();
-    expect(deliveryModes).toEqual([{ code: 'mode1' }, { code: 'mode2' }]);
-  });
-
-  it('should be able to get selected delivery mode', () => {
-    store.dispatch(
-      new fromCheckout.LoadSupportedDeliveryModesSuccess([
-        { code: 'mode1' },
-        { code: 'mode2' },
-      ])
-    );
-    store.dispatch(new fromCheckout.SetDeliveryModeSuccess('mode1'));
-
-    let selectedMode: DeliveryMode;
-    service.getSelectedDeliveryMode().subscribe(data => {
-      selectedMode = data;
-    });
-    expect(selectedMode).toEqual({ code: 'mode1' });
-  });
-
-  it('should be able to get the code of selected delivery mode', () => {
-    store.dispatch(
-      new fromCheckout.LoadSupportedDeliveryModesSuccess([
-        { code: 'mode1' },
-        { code: 'mode2' },
-      ])
-    );
-    store.dispatch(new fromCheckout.SetDeliveryModeSuccess('mode1'));
-
-    let selectedModeCode: string;
-    service.getSelectedDeliveryModeCode().subscribe(data => {
-      selectedModeCode = data;
-    });
-    expect(selectedModeCode).toEqual('mode1');
-  });
-
-  it('should be able to get the card types', () => {
-    store.dispatch(
-      new fromCheckout.LoadCardTypesSuccess([
-        { code: 'visa', name: 'visa' },
-        { code: 'masterCard', name: 'masterCard' },
-      ])
-    );
-
-    let cardTypes: CardType[];
-    service.getCardTypes().subscribe(data => {
-      cardTypes = data;
-    });
-    expect(cardTypes).toEqual([
-      { code: 'visa', name: 'visa' },
-      { code: 'masterCard', name: 'masterCard' },
-    ]);
-  });
-
-  it('should be able to get the delivery address', () => {
-    store.dispatch(new fromCheckout.SetDeliveryAddressSuccess(address));
-
-    let deliveryAddress: Address;
-    service
-      .getDeliveryAddress()
-      .subscribe(data => {
-        deliveryAddress = data;
-      })
-      .unsubscribe();
-    expect(deliveryAddress).toEqual(address);
-  });
-
-  it('should be able to get the address verification result', () => {
-    store.dispatch(
-      new fromCheckout.VerifyAddressSuccess({ decision: 'DECLINE' })
-    );
-
-    let result: AddressValidation | string;
-    service
-      .getAddressVerificationResults()
-      .subscribe(data => {
-        result = data;
-      })
-      .unsubscribe();
-    expect(result).toEqual({ decision: 'DECLINE' });
-  });
-
-  it('should be able to get the payment details', () => {
-    store.dispatch(new fromCheckout.SetPaymentDetailsSuccess(paymentDetails));
-
-    let tempPaymentDetails: PaymentDetails;
-    service
-      .getPaymentDetails()
-      .subscribe(data => {
-        tempPaymentDetails = data;
-      })
-      .unsubscribe();
-    expect(tempPaymentDetails).toEqual(paymentDetails);
-  });
-
   it('should be able to get the order details', () => {
     store.dispatch(new fromCheckout.PlaceOrderSuccess({ code: 'testOrder' }));
 
@@ -173,73 +49,6 @@ describe('CheckoutService', () => {
     expect(orderDetails).toEqual({ code: 'testOrder' });
   });
 
-  it('should be able to create and set address to cart', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    service.createAndSetAddress(address);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.AddDeliveryAddress({
-        userId: userId,
-        cartId: cart.code,
-        address: address,
-      })
-    );
-  });
-
-  it('should be able to load the supported delivery modes', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    service.loadSupportedDeliveryModes();
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.LoadSupportedDeliveryModes({
-        userId: userId,
-        cartId: cart.code,
-      })
-    );
-  });
-
-  it('should be able to set the delivery mode', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    const modeId = 'testId';
-    service.setDeliveryMode(modeId);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.SetDeliveryMode({
-        userId: userId,
-        cartId: cart.code,
-        selectedModeId: modeId,
-      })
-    );
-  });
-
-  it('should be able to load supported cart types', () => {
-    service.loadSupportedCardTypes();
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.LoadCardTypes()
-    );
-  });
-
-  it('should be able to create payment details', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    service.createPaymentDetails(paymentDetails);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.CreatePaymentDetails({
-        userId: userId,
-        cartId: cart.code,
-        paymentDetails,
-      })
-    );
-  });
-
   it('should be able to place order', () => {
     cartData.userId = userId;
     cartData.cart = cart;
@@ -251,56 +60,6 @@ describe('CheckoutService', () => {
         userId: userId,
         cartId: cart.code,
       })
-    );
-  });
-
-  it('should load address verification results', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    service.verifyAddress(address);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.VerifyAddress({
-        userId: userId,
-        address,
-      })
-    );
-  });
-
-  it('should set delivery address', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-    service.setDeliveryAddress(address);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.SetDeliveryAddress({
-        userId: userId,
-        cartId: cartData.cart.code,
-        address: address,
-      })
-    );
-  });
-
-  it('should set payment details', () => {
-    cartData.userId = userId;
-    cartData.cart = cart;
-
-    service.setPaymentDetails(paymentDetails);
-
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.SetPaymentDetails({
-        userId: userId,
-        cartId: cartData.cart.code,
-        paymentDetails,
-      })
-    );
-  });
-
-  it('should be able to clear address verification result', () => {
-    service.clearAddressVerificationResults();
-    expect(store.dispatch).toHaveBeenCalledWith(
-      new fromCheckout.ClearAddressVerificationResults()
     );
   });
